requirements.txt
fastapi
uvicorn
openai
python-multipart
PyMuPDF
python-dotenv
streamlit
requests
beautifulsoup4==4.12.3
datetime
fpdf2
Pillow
python-docx
stripe
ngrok
yagmail